mv loading component and rename to Spinner.vue
