+++ src/components/AdminAccountList.vue
@@ -33,7 +33,7 @@
 
 <script lang="ts">
     import AccountList from "./AccountList.vue"
-    import LoadingComponent from "../components/LoadingComponent.vue"
+    import LoadingComponent from "./loading/Spinner.vue"
     import { ref } from 'vue'
  
     export default {
