+++ src/components/loading/Spinner.vue
@@ -1,7 +1,7 @@
 <template>
     <div class="w-full mt-10 items-center justify-center flex flex-col">
         <h2 class="text-xl text-gray-800">Loading Contents...</h2>
-        <img src="../assets/loading-spinner.svg" class="w-1/12"/>
+        <img src="../../assets/loading-spinner.svg" class="w-1/12"/>
     </div>
 </template>
