+++ src/views/lecturer/CourseFormSuspenseWrapper.vue
@@ -12,7 +12,7 @@
 
 <script lang="ts">
 import LecturerCreateCourseForm from "./EditCreateCourseForm.vue";
-import LoadingComponent from "../../components/LoadingComponent.vue"
+import LoadingComponent from "../../components/loading/Spinner.vue"
 import { store } from '@/store/store';
 import {Role} from '@/entities/Role'
 import router from '../../router';
